sendSlackMessage: don't send again if the there it increased by one finding every time
use filteredMessage as a way to prevent over notification on slack.
This would not send again any notification if there is an extra finding on a given PR,
for the previous 50 message, or a day ago.

This would debounce chaingun messages in any slack channel from the main action itself.

It should not have any other side-effect.

This will effectively change the hash function, so invalidate the debounce for all the previous messages.