Reduce alerting when MSK_FETCHER_PASSWORD value is used
This alert appears quite commonly within the search repository. In
some cases, it's the only alert that appears too. Given this appears
to be a safe pattern, it makes sense to reduce alerting when this is
used without turning off any of the other alerting that should occur
from this rule.