Allow rules to specify specific repositories to ignore or run in
This is useful in the case where we want to make specific modifications
to rules for only a select number of repositories. One example of this
is that we want to allow expect usage, but still alert on unwrap in
Search specific repositories. However, within Brave-core we want to
make sure we alert on both. This should allow us to fork rules to
make them specific to the repository they run on.